+++ src/test/java/ch/saunah/saunahbackend/service/MailServiceTest.java
@@ -34,7 +34,7 @@ class MailServiceTest {
         user.setEmail("test@mail.ch");
         String verificationId = "1";
         int userID= 1111;
-        double resetToken = 12345;
+        int resetToken = 12345;
         assertDoesNotThrow(() -> mailService.sendUserActivationMail(user.getEmail(), verificationId));
         assertDoesNotThrow(() ->mailService.sendPasswordResetMail(user.getEmail(), userID ,resetToken));
         user.setEmail("bad email");
